Update tests to accommodate `employee_id` as lookup field
Previously, the tests assumed the `id` field of a Technician and
Salesperson was the field being used to create an appointment, delete
a Salesperson, and delete a Technician. However, it is possible for
the `employee_id` field to be used for these operations instead (as
this requirement is not specified in the project requirements). This
commit updates the tests to account for both possibility of `id` and
`employee_id` being used as the lookup field.

Additionally, a bug was fixed involving reference of an undefined
variable in the `test_technician_delete` test.